requirements.txt
fastapi
speechrecognition
google-generativeai
gtts
pygame
uvicorn 
pydantic-settings
python-dotenv 
sounddevice
soundfile 